requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.3
apscheduler==3.10.4
python-dotenv==1.0.1
sendgrid==6.11.0